[FIX] account_reports: fix broken customer statements with cron
Steps to reproduce:
- Accounting > Reporting > Partner Ledger
- Select multiple partners
- Select "Customer Statements" as report
- Click "Send"
- Go to the one of the partner in question an dopen the generated PDF
- The address is not the one of the company, the footer doesn't show the right company, etc

Cause:
When sending to multiple partners, the action is dispatched with cron so a new `env` is created. This `env` has all companies active and the first one as main company.
So in the end the wrong company is used to generate the report.

Solution:
In the cron when calling `_process_send_and_print` we give `report.with_company(company)` to make sure that the value the report will us are the one from the correct company.

Also revert the small [fix](https://github.com/odoo/enterprise/pull/82320/files) that was made before, which is no longer needed.

opw-4635283

closes odoo/enterprise#84482

X-original-commit: 330d3ca826877f45a57e78edf3aa2c2759a01fd2
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>