[FIX] account_reports: fix tests in partner_ledger & reconciliation report
This commit fix some tests in account_reports which were failing in nightly
on the runbot, see error 134654.

First test fails because an account number sequence was changing depending
on if the accountant module was installed or not. Fix the test by replacing
the hardcoded account name by the used bank journal name.

Second test fails because of payment_state computation which is different
depending on if the accountant module was installed or not. Fix it by using
the _get_invoice_in_payment_state method instead of hardcoded value.

no task

closes odoo/enterprise#85621

X-original-commit: b9ce9799446bb4fc0d411b4215801f48fd6660c1
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Igor Bertrand (igbe) <igbe@odoo.com>