[FIX] sale_stock_renting:qty_delivered
Fix qty_delivered when a rental order is processed through the PoS. This
commit ensure that the PoS qty_delivered computation is done after the
other calculation

Steps to reproduce:
-------------------
* Make a rental for 1 product
* Open the sale order in PoS
* Pay for the rental
> Observation: The qty_delivered is set to 2 instead of 1 on the rental

Why the fix:
------------
When the order is a rental processed through the PoS, we recompute the
rental qty and add the PoS quantity to make sure that the value is
correct. The computation is actually just the fusion of the PoS and
rental _compute_qty_delivered methods.

opw-4582505

closes odoo/enterprise#86774

X-original-commit: 189edebbc8faaca53b25d8b82386fb9b91b74be1
Signed-off-by: Robin Engels (roen) <roen@odoo.com>