[FIX] ai: use O2M for attachments instead of M2M
Using a `Many2many` for the `attachment_ids` and `url_attachment_ids`
fields was problematic.
The `ir.attachment` model is meant to be used as a `Many2one` or
`One2Many` since it is linked to its parent record using ressource
fields (`res_model` and `res_id`).
When using a `Many2many` field, the `res_id` loses its meaning since it
should be able to include multiple IDs to be accurate.

In practice, this caused an access rights issue because the
`ir.attachment` model [1] relies on the `res_id` to check which user is
allowed to access the attachment.

To fix this, the attachment fields of the agent are replaced with
`One2many` fields. This implies that the attachments aren't shared
anymore and the embedding will need to be recomputed in case of
duplicated attachments across multiple agents.

[1] See function `check` of `base/models/ir_attachments.py`

closes odoo/enterprise#84427

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>