[FIX] account_reports: fix hidden cog button in responsive
Steps to reproduce:
- Install Accounting
- Install a module that adds a button to tax report (e.g. l10n_ar)

- Switch to an Argentinian company (e.g. (AR) Responsable Inscripto)
- Go to "Accounting / Reporting / Statements Reports / Tax Return" (i.e. Tax Report)
- Select "Argentinean VAT Book (AR)" as report
- A cog button should appear after the report name with "VAT Book (ZIP)" option

- With browser debug feature, switch to a responsive display and reduce the width
or
- Repeatedly increase the zoom to increase the size of the button

Issue:
For a certain range of width the cog button is not visible.

Cause:
".o_control_panel_breadcrumbs" element that contains the buttons has a min-width
of 200px, hiding all elements that are after these 200px when reducing the width.

Solution:
Increase the min-width to 350px.
The value is arbitrary but it covers well this use case.
With a lower value, the cog button could still be hidden in the case of the
Argentinian Tax Report in Spanish language.

opw-4747697

closes odoo/enterprise#85055

X-original-commit: 7f7e274d0ad4731ee077cd6e793f08972ef80c87
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>