[IMP] web_enterprise: set global text-color to `900`
The goal is to increase contrast between the global text-color and
text-muted in the backend while keeping the current text-muted color.

This is done because the contrast was too small between these two
colors.

The contrasting color for headings and active components is set to
`$black` (#000 and #FFF darkmode).

To avoid the secondary buttons looking "disabled" the color is also
adapted.

task-4756587

Part-of: odoo/enterprise#85027
Related: odoo/odoo#209125
Signed-off-by: Pierre Paridans (app) <app@odoo.com>