[IMP] l10n_ec_edi: add portal withholding view and fix printing
For Ecuador, Purchase Withholdings are a common document that needs
to be visible to vendors to make sure everything is as expected on both
sides.

Since they are of type entry, they aren't currently visible on the
portal view so a new section needed to be added. It is implemented as
another filter on the invoicing portal controller and slightly modified
views.

As a result of adding the portal view, the previous implementation of
blocking the print button with a warning saying to use the context
action instead is no longer working. As such, the print button has been
implemented to display the right report as well as the account.move.send
wizard being able to be used to match invoices and vendor bills.

This allows for the invoice report id to be properly set and allow us to
download the edi documents automatically in the portal.

task-4304037

closes odoo/enterprise#83178

Related: odoo/odoo#205420
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>