[IMP] ai: combined ai draft and agents in one module
As the title suggest the following commit took combined the work done
in the AI agent's pr as well as the current work done for the new text
draft ai integration and put them inside the same module.

Currently, they still work differently - thus some things inside the
module might be a bit ugly / messy. In the future, the two will also
share the same backbone - thus the code will ultimately look cleaner.

Also fixed a tiny bug with the text select. We were not passing a model
which is necessary because the "initialise context" is a model method now.

task-4526290

closes odoo/enterprise#85028

X-original-commit: 1a6de6cc123f162f3602e11a53db2396ae18327f
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>