[FIX] l10n_in_reports_gstr: avoid already reconciled bill past bill
In GSTR-2B, while doing the matching for previous bills, we want to avoid
The bills that are already reconciled and marked as partially matched,
fully matched or manual matched.
Sometimes, it's possible that there can be more than one bill with the same
The reference number and  if any of them has been already reconciled then
We want to avoid it again being reconciled in next return periods.

opw-4744438

closes odoo/enterprise#84721

X-original-commit: fa797e6c2056d3db0e2ce248a2a99c7d733c172e
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Vivek Pathak (vivp) <vivp@odoo.com>