[IMP] l10n_es_reports: leave Mod347 field empty on bill creation
If a vendor bill is created by the expense app for an expense report
paid by an employee, we force the Modelo 347 field to be empty.
Reimbursement of expense to employees are not considered third-party
transactions in this context, as they are part of the company's internal
employment relationship.
We make the improvement in the l10n_es_reports module because there is
no bridge module between es and expense, this won't change the standard
flow : it will work whether the hr_expense is installed or not.

closes odoo/enterprise#85353

Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>