[IMP] planning: display open shift in the front-end
This commit's purpose is to have a better visual of the open shifts in
the planning app for external user. Instead of having the list with
values, the open shifts are now displayed in the calendar view with a
specific background color to be able to differentiate them from
employee shifts at one glance.

closes odoo/enterprise#85121

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>