[FIX] l10n_be_intrastat: error when exporting xml if no product
How to reproduce:
- Select a Belgian company
- Create an invoice for a EU country
- Add a line with no product
- Display the intrastat report for the invoice period
- Export the xml

The default sql query allows the selection of account move lines with no product to be able to trigger a warning. Those lines are then filtered out when displayed in the web interface. Those lines were not filtered out when getting data for file exports. An error would trigger rendering the xml as the weight would be None where an integer is required.

This commit adapts the product_type_condition to exclude lines without products when a file is exported.

Task 4763185
See opw-4671512, opw-4735439, opw-4642781

closes odoo/enterprise#84849

X-original-commit: d087b120ec7befe96908ce20f42626d841b76d89
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>