[IMP] iot: ui/ux improvements
We added back the amount of devices connected to an IoT Box on the IoT Box card.
We also added the pairing date on the card, for a user to spot faster an IoT Box.

Some elements have been moved, and some sentences have been rephrased on the wizards to pair an IoT Box.

closes odoo/enterprise#85320

Task: 4585446
X-original-commit: 2160f4d86563f059e10758aaa62e681d1240a8e3
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>