[IMP] iot: ui/ux improvements
We added back the amount of devices connected to an IoT Box on the IoT Box card.
We also added the pairing date on the card, for a user to spot faster an IoT Box.

If an IoT Box is linked to a PoS, the PoS configs where devices are linked are now listed on the IoT Box
cards.

Some elements have been moved, and some sentences have been rephrased on the wizards to pair an IoT Box.

closes odoo/enterprise#84591

Task: 4585446
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>