[FIX] l10n_br_avatax: stop limiting description length
In some cases, it's not possible for customers to trim the description
of an invoice line to the accepted length (e.g. for traceability or
completeness).

It turns out that Avalara will trim and transform the descriptions on
our behalf to make them compatible with the government's API. As part
of that process, they translate `\n` characters to ` | ` before
sending it to the government.

Unfortunately, there seems to be a bug with that process on their
side. When a '\n' character is included in the description, they don't
trim the description they send to the government sufficiently. Perhaps
they trim before the '\n' to ' | ' translation, or perhaps they have
some other bug.

We sidestep the issue by doing that same translation on our side and sending
them that. It doesn't result in any change on the EDI PDF or XML, and
this way we can avoid hardcoding these limits on our side.

task-4726372

closes odoo/enterprise#85080

X-original-commit: 3ebb91c1deb9ef309139d1f95814f788b3d567e4
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>