[FIX] quality_mrp: avoid QC reassignment when splitting productions
Steps to reproduce:

- Install both mrp and quality_control
- Create 2 products:
    - FP (Final Product)
    - COMP (Component) tracked by SN > pu SN1 to SN4 in stock
- Create a BOM for FP using COMP as component, add an operation with
  an instruction of type "Register Consumed Materials" for SN control
  per product.
- Create and conform an MO for 2 units of FP: SN1 and SN2 are reserved
- Click on the wheel icon > Split production > Split in 2
- On the MO-001, SN1 is reserved
- Open the operation in the shopfloor and proceed with the quality check
> It is currently refering to SN2 !!
- register SN3
> It did not alter the value of the lot referred to on MO-001.
> It did update the serial number present on MO-002 from SN2 to SN3.

Cause of the issue:

Confirming the MO will in turn confirm the workorder and create a
quality check for the consumed material registration of SN:
https://github.com/odoo/enterprise/blob/a26df8b50b923a2b23b8406d798febbdf1a72415/mrp_workorder/models/mrp_workorder.py#L625-L628
https://github.com/odoo/enterprise/blob/a26df8b50b923a2b23b8406d798febbdf1a72415/mrp_workorder/models/mrp_workorder.py#L415-L421
At this point, the create quality check is linked to the MO but also
to the first move line (sml1) of the raw move (m1) refering to SN in
that MO:
https://github.com/odoo/enterprise/blob/a26df8b50b923a2b23b8406d798febbdf1a72415/mrp_workorder/models/mrp_workorder.py#L495-L503
However, the `move_id`, `move_line_id` and `production_id` are all three
stored non computed fields of the `quality.check` model. The issue will
then be that when the production is split, MO will be renamed MO-001
the QC will still refer to MO-001, sm1, sml1 but sml1 will be reassigned
to the raw move of the new production MO-002. In particular, while
the QC is still accessed from MO-001 and still refer to the raw move sm1
of  MO-001, it actually updates the values MO-002 since performing these
kind of quality checks actually updates the qty and lots on the move
line it is related to:
https://github.com/odoo/enterprise/blob/a26df8b50b923a2b23b8406d798febbdf1a72415/mrp_workorder/models/quality.py#L507-L508
Now, the reason why sml1 is reassigned to the mo to back order rather
than staying on sm1 and hence on MO-001 is because the order of stock
move line is altered to be reassigned by these lines during the
`_split_production`:
https://github.com/odoo/odoo/blob/e4bd46930d1841fcd50bf426f5625e9495213ef7/addons/mrp/models/mrp_production.py#L1890-L1895
Currently, this order puts sml with an existing quality check after
stock move lines without QC:
https://github.com/odoo/enterprise/blob/a26df8b50b923a2b23b8406d798febbdf1a72415/quality_mrp/models/stock_move_line.py#L36-L41
As such, they will end up being assigned with the latest backorder to
assign.

Fix:

Since these kind of QC's are created at confirmation and since the first
MO will not be reconfirmed, it is important that the QC created for
MO-001 stays linked to that MO. In order to achieve this, move lines
associated with a QC should definitely stay assigned to the first MO
and the sml should therefore be ordered by these without QC last.

opw-4656240

closes odoo/enterprise#87701

X-original-commit: 4eb4fb518626216b2e2e922d2f93f378c12a19d9
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>