[FIX] l10n_pe_edi_pos: refund orders in pe PoS
**Problem**
When purchasing items, then refunding them in PoS with a Peruvian
company, a traceback appears, making the refund impossible.

**Steps to reproduce**
- Install "l10n_pe_edi_pos"
- Change to the Peruvian Company
- Create an order from PoS, ask for an invoice
- Create a refund for that order
- A traceback appears

**Why the fix**
A previous fix (8af3821) was done to make sure we could select
multiple orders in the case of a global invoice. This fix did
not take into account the fact that the *l10n_pe_edi_refund_reason*
could be not set, and get it's default value being False. So when
trying to use the join() method on a boolean instead of a string,
the traceback appeared. With this fix, we only use the join method
if the *l10n_pe_edi_refund_reason* is set, and keeping the join()
method to handle the case where multiple *l10n_pe_edi_refund_reason*
are set.

opw-4761685

closes odoo/enterprise#85543

X-original-commit: 9e7abc1dddf21e924b917c2d622a82d29bd2b08a
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Arthur Nanson (artn) <artn@odoo.com>