[FIX] sign: prevent signing with whitespace-only name in auto mode
Currently, entering only spaces in the Full Name field when signing in
auto mode enables the sign button, allowing users to proceed
without a valid name.

Steps to reproduce:
1. Sign > Create a request
2. Attempt to sign in auto mode with only spaces in the Full Name field
3. Observe that the sign button becomes enabled

Cause:
The system does not checks for whitespace-only name
Solution:
Add a validation to disable the sign button if the trimmed name is empty in
auto mode.

opw-4628577

closes odoo/enterprise#86225

X-original-commit: 83f149eae3436ac48f0ac719cf6c8c3c5894d341
Signed-off-by: Piyush Sharma (pish) <pish@odoo.com>