[IMP] industry_fsm,project_sale_subscription,documents_project: clean project settings
After this commit, we remove the FSM global setting "Extra Quotations" to only keep it in the project form view.
The way we handle the group "industry_fsm.group_fsm_quotation_from_task" is similar to what we do in this PR: https://github.com/odoo/odoo/pull/208005
Also, as it is already the case for "Milestones" and "Task Dependencies" we do not enable the setting "Recurring Tasks" on new FSM projects.
We also make some adjustments in recurrence with sale subscription as we removed the global setting.

task-4700865

closes odoo/enterprise#84657

Related: odoo/odoo#208005
Related: odoo/upgrade#7640
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>