[FIX] l10n_us_check_printing: align check format with the expected
Steps to print the check:
1. Ensure l10n_us_check_printing is installed
2. Set the check layout to Print Check (Middle) - US
3. Create a PO and a bill from it.
4. Create a payment for the bill using checks as your payment method
5. Click Print Check on the payment

The printed check will now match the format and allignment of the
templates used here:
https://checkdepot.net/collections/odoo-checks/products/odoo-checks-top-
format. Additionally, the check will now avoid printing the issuing
company name on the stubs when the manual numbering setting on the bank
journal is switched off. This assumes that the checks are preprinted
with the company name and check number.
Before the fix, the checks would print the vendor's name in the stubs
and the date allignment in the (middle) format was overlapping with the
check number.
After the fix, the checks are alligned with the expected format and
conditionally renders the correct fields in the document's intended
areas.
opw-4557006
opw-4738359

closes odoo/enterprise#85916

X-original-commit: afa85a637959f64dfdc9f24a0c7ae93fb2588bd1
Related: odoo/odoo#210653
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Ezzuldin Zaky (ezza) <ezza@odoo.com>