[FIX] voip: correct handler for ontouchend
The touchend event must be preventDefault-ed to prevent zooming on
double-tab on mobile. However, the touchend handler was incorrectly set
to this.props.onClickKey, which doesn't exist.

This commit updates the code to trigger a click event on touchend so
that the same handler is used whether the button is tapped or clicked,
preventing similar situations where the click handler is updated but not
the touchend handler, leaving the latter broken.

closes odoo/enterprise#84740

X-original-commit: c3e809a807a3704e8e418f6232a7e1cfb9e8c3e1
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>