[FIX] account_reports: date filter styling
For languages with longer terms than the English ones, the labels in the
date filter would overflow because they had a fixed width. We now change
it to use a minimal width and expand with the content.

Also, the quarter label was not translatable in other languages that do
not use the notation "Q1 2025". It's made translatable now.

Lastly, some padding caused custom date inputs to be too narrow, meaning
a horizontal scroll was necessary to see the whole value. This was
fixed as well.

task-4770592

closes odoo/enterprise#86557

X-original-commit: 2f3180d1a4b42f58cd17797194e8b6ffbdd67685
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>