[FIX] l10n_lu_reports: use custom "from date" in P&L export
Steps to reproduce:
- Install Accounting and l10n_lu_reports
- Switch to a Luxembourgish company (e.g. LU Company)
- Go to "Accounting / Reporting / Statement  Reports / Profit and Loss"
- Set custom date from "01/01/2025" to "31/12/2025"
- Export XML

Issue:
In the exported XML, for "CA_BILAN" declaration section, the start date
is set to "01/12/2025" instead of "01/01/2025".

Cause:
When generating the report values of that section, the options are
recomputed with the comparison filter.
This section is based on "Balance Sheet" report and recomputing the
options of that report provides a different "from date" coming from
"_init_options_date" method.

Solution:
Force the "from date" of the original options when using a custom range
date.

opw-4737810

closes odoo/enterprise#85218

X-original-commit: 0cdac6a48eeba904e672816f0e81e73c7ef54119
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>