[FIX] documents: rectify detailspanel selection
- The details panel does not update when clearing the selection through the
  control panel.

Steps to Reproduce:
====================
- Open the detailPanel.
- Select a document.
- Click on a blank space, which removes the current selection.
- Re-select a document.
- Click the ‘x’ button on the control panel to clear the selection.
- The document selection in the detailPanel does not get cleared.

Technical
===========
In DocumentsRightPanel and how DocumentsDetailsPanel is handled, The panel
relies on the `state.focusedRecord` to determine which record's details to
display. However, when clicking the ‘X’ button in the control panel to clear the
selection, the `state.focusedRecord` is not being Updated or cleared, the detail
panel still shows the previously focused record.

After this commit:
Clicking the ‘x’ will properly clear the selection from the detailPanel.

Task-4752944

closes odoo/enterprise#91275

X-original-commit: 44da2bfd72177903e05910abdc20fc128b05925d
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>