[IMP] account_accountant: apply full amount widget
This commit will add the possibility to quickly apply a full amount when you
have a partial amount.
Also removing part of the edit line function that is not needed anymore since we
added the amount currency in the view, the record_data parameter has now always
the amount_currency in it.

And doing some formatting because the alignment was wrong in the form view.

task-4749342

closes odoo/enterprise#85838

X-original-commit: bfa461ec9522089c9a4286cd30ce287b843e2078
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>