[FIX] quality_control: ensure quality check creation for MOs
Steps to reproduce the bug:
- Create a quality point with the following settings:
    - Measure on: Operation
    - Picking Type: Manufacturing
    - Product Category: "All"

- Create a storable product “P1”:
    - Product Category: "All"
    - BoM:
         - components:
             - 1 unit of P1

- Create a manufacturing order to produce one unit of P1
- Confirm it

Problem:
The manufacturing order is confirmed, but the corresponding quality
check is not created.

This issue occurs when a quality point is configured with "Measure on:
Operation". In that case, the quality check should be created for
manufacturing orders here:
https://github.com/odoo/enterprise/blob/18.0/quality_mrp/models/stock_move.py#L49-L51

However, an empty record is passed for the product parameter.
As a result, the domain defined here:
https://github.com/odoo/enterprise/blob/18.0/quality_mrp/models/stock_move.py#L19-L21

is evaluated with both product and category set to False, which
prevents the created quality point from being matched and thus no
quality check is created.

https://github.com/odoo/enterprise/blob/6ee3472937118e758399a9577251efad8c4c1195/quality_control/models/quality.py#L175-L177

opw-4762914

closes odoo/enterprise#85383

X-original-commit: 678596981d5e489ccf0125d83647aeb0bb78ceb1
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>