[IMP] l10n_uk_reports: send module version
Fixes were made regarding the fraud prevention headers, but we have
still odoo requests that send invalid headers.

When checking the odoo requests on the hmrc website, we can group by
`Gov-Vendor-Version`, which is currently set to the Installed Version of
the base module, twice (for some reasons).

This commit send the version of the `l10n_uk_reports` module.  So when
a new fix is made, we can bump the module version and see if failing
requests are up-to-date.

task-4627086

closes odoo/enterprise#86354

X-original-commit: 3b263f4f9044126d2abac418f8fa1d6844940eaa
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Arnaud Sibille (arsi) <arsi@odoo.com>