[FIX] documents_hr_contract: add fallback user for document owner
It is possible that a employee won't have a contract (maybe a
freelancer) nor a user related. In such case when asked to sign
a document, the document cannot be signed. It is possible to
reproduce this behaviour on runbot

This commit aims to set a user, since the user cannot be false.
I believe that since one of the options would be set it to the
manager of the contract the employee is under, it would be a valid
option considering that the employee has no contract, to set it to
whoever sent the request, in this case, the request create_uid

opw-4750010

closes odoo/enterprise#85330

X-original-commit: eadd2fa92699b577b34f8494c629425abd42afd9
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>
Signed-off-by: Lucas Rizzi Falasca Duarte (lrfd) <lrfd@odoo.com>