[FIX] documents: create documents with correct company
Creating documents through a mail alias for a folder that was not the
default company on the database would cause an error. This was due to
the company on the inactive alias that was being created for this new
document not matching the company of the owner document.

Added the company id to the vals_list before the creation of the
document in order to use the proper company on the alias.

opw-4717277

closes odoo/enterprise#87817

X-original-commit: 9316b50726fc6a049a9c2df0de556ebc5b3298e6
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Ryan Cen (ryce) <ryce@odoo.com>