[IMP] sign, test_mail: Activity revamp finishing touches
task-4592571 changed up the way activities function, but not everything
was completed before the end of the freeze; in addition, it included
a number of bugs/crashes/usage inconveniences.

This commit adds:
- Sign "Request signature" activity type renamed to "Signature"

This commit also fixes:
- Postfreeze checkout comment on test_mail_enterprise query count
change deleted; as action_feedback behavior changes, it can be expected
that query count would change.

task-4747156

closes odoo/enterprise#85824

X-original-commit: 2f348bbbd3bccda261f5529c15a8a8adcfc5f66e
Related: odoo/odoo#210497
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>