[FIX] account_intrastat: report unfold lines with company incoterm
The incoterm code of a line in the intrastat report is the incoterm code from the move or if there is no incoterm code on the move, the one from the company.
When creating the request when unfolding report lines, the incoterm code was always checked on the move and not on the company. This would not return lines without incoterm code.
This commit changes the incoterm domain to also check the incoterm code of the move company when no incoterm code is set on the move.

Steps to reproduce:
- From an EU company, create an invoice to a company in another EU country
- Do not set an incoterm on the invoice
- In settings, set a default incoterm on the company
- Display the intrastat report for the time period of the invoice
- Try to unfold the line corresponding to the invoice

opw-4642781

closes odoo/enterprise#85098

X-original-commit: 2e14f318b5905655c088d4dcf6f5b9b861acca15
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>