[FIX] quality_control: trigger quality_check for unchecked moves
Steps to reproduce:

- Create 2 storable products: Product A, Product B
- Create a control point pass/fail control by quantity on receipt and
  set both products on it.
- Create and confirm a receipt with 2 moves:
    - 2 x Product A
    - 2 x Product B
- Click on the "Quality Check" button and mark one unit of Product A
  as failed.
> This should open the QC for product B.
- Close the QC for prodcut B wihtout setting it as failed or passed.
- Click on the "Quality Check" button
> Nothing happends

Expected behavior:

The QC for product B should open.

Cause of the issue:

When the quality check is marked as failed, the related move is picked:
https://github.com/odoo/enterprise/blob/679c81924e05e7544eecfdb960c9ed9611e9cdd5/quality_control/models/quality.py#L438-L451
However, the move for your second porduct is not and hence, its product
is not considered to be checkable by the `check_quality` action:
https://github.com/odoo/enterprise/blob/679c81924e05e7544eecfdb960c9ed9611e9cdd5/quality_control/models/stock_picking.py#L61-L69
Since these lines were added in c83728f3a7f1504b272f4ef97e4187868338adb3
to not bypass the QC of the unpicked move at validation we should
probably only add that filtering condition in that case and continue to
consider the unchecked moves as checkable prior to validation.

opw-4765075

closes odoo/enterprise#88122

X-original-commit: 3052bdadcb25ac818e0827d17117bd829ab58bac
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>