[IMP] ai: implementation of a vector field
This commit introduce a new `Vector` field for the AI module.

Before, the module was defining the vector column by manually adding the
sql queries (see `ai.embedding` module). This commit solves that issue
by introducing a new orm field `Vector` which takes care of the creation
and read of columns of type vector.

task-4763397

closes odoo/enterprise#84748

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>