[FIX] l10n_nl_reports: fix error description for SBR status service
Commit baa934898f184b10ae4c5c51e26fabe4e0178a86 updated the body for ongoing
process responses when checking the status of a report submitted through SBR.

The variable `error_description` it tries to assign is not defined in that
scope of the function, resulting in an error.

This PR reverts the value assigned to be the status description, as it was
prior to the above commit.

opw-4775345

closes odoo/enterprise#85063

X-original-commit: 8133de3320831af73c6e1df179246d5673e07bd3
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Florent Huylenbroeck (flhu) <flhu@odoo.com>