[FIX] account_reports: Avoid redirection when reviewing all checks
Before the commit:
On a return which passed all checks in 'new' state, when we clicked the dots then click review all checks.
It would open the checks view with all the checks green then instantly redirect to the return view.
It would also automatically forward to the next state.

This is because there is a call to refresh the checks on the page load of the checks view.
During this page load, a call to try_forward_state was triggered even if we didn't want to change state.
Now, when we open the view from the button 'review all checks', we would not allow changing automatically to the next state.
The only way is via the 'Review Checks' main button.

closes odoo/enterprise#85222

X-original-commit: 6ad093930196070bf328dee34e192b66f560cfa6
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>