[FIX] mrp_workorder: set correct employee on work order via Shop Floor
Task: 4610586

If an operator is added in the side panel of Shop Floor, and an MO has
multiple work orders, then after marking all of the steps as done, the
selected user is assigned to all of them except the final one.

This commit fixes the error, assigning the selected operator to all
completed work orders.

closes odoo/enterprise#88695

X-original-commit: 0db820004b0ac6f7a1bff938f5ee52de62580743
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Pawel Fertyk (pafe) <pafe@odoo.com>