[FIX] account_accountant: fix getting amount currency from transaction
The Issue:
Prior to this commit, retrieving the amount currency from the transaction details caused a traceback. This occurred because the transaction details are stored as a jsonb object, and the code attempted to access a string within a dictionary, resulting in a TypeError mismatch.

The Fix:
The jsonb object is now converted to a string before processing. Additionally, an IndexError is caught to handle cases where the expected group is not found in the match.

opw-4754518

closes odoo/enterprise#85011

X-original-commit: 96f5915a492fc51b79ce25cd5db404bad099abd4
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>