[FIX] documents_account_peppol: fix reception of documents
When receiving Peppol invoices through the Documents App, a silent
error is raised. This comes from previous change [1] of the method
return type. We changed it from boolean to `account.move` and did
not changed it in the documents override causing an append
of `account.move` with boolean which crashes.

[1]: https://github.com/odoo/odoo/pull/197233
task-no (PMAX testing)

closes odoo/enterprise#85167

X-original-commit: 4a1e8782a0f276217f6f42444ffbbdb0a57cbd0b
Related: odoo/odoo#209320
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>