[FIX] industry_fsm: allow timesheet user to start timer on task
Before this commit, when the project is private and the user can only
access to some tasks inside that project, he can create timesheets but
cannot stop timer in one of those tasks when `industry_fsm` module is
installed.

This commit makes sure the user can start/stop a timer on task for which
he has access to even if the project is private and he does not have
access to it.

Steps to reproduce the issue:
----------------------------

0. install industry_fsm module
1. Create a project A with "followers" as privacy visibility
2. Create a task A and assign it to a project user with lowest timesheet
   access right.
3. log in as that user
4. Start a timer on Task A
5. Stop the timer and confirm the wizard

Expected Behavior:
-----------------

The timesheet should be created without any issue

Current Behavior:
----------------

An Access error is raised because the user has no access to the project.

task-4680412

closes odoo/enterprise#91546

X-original-commit: 49f880c745918c312382a28dcefafadb02139193
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>