[FIX] account_budget: handle 0 purchase order line with negative invoiced qty
It's possible to have a 0 purchase order line with a negative invoiced qty if
there is only a credit note linled to the PO. When the budget report process
this order line, it will trigger a traceback.

This commit aims to handle this case so the report can be open without issue.

opw-4749480

closes odoo/enterprise#85056

X-original-commit: 220fe704929217575ed9f8880649c97e3b3ef1be
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Thomas Beckers (tbs) <tbs@odoo.com>