[FIX] whatsapp: resend works for channel messages
Currently the "resend" mechanism relies on `_send`;
marking messages to resend as "outgoing".

However when used in batch mode `_send` only sends template messages.

We enable `_send_cron` to send discuss messages too and to process these
in priority. As these have a fixed time limit to be sent.

We also do some filtering on resend to avoid marking messages
that we know will fail as outgoing. Including "unrecoverable" errors
and messages on outdated channel.

Messages on outdated channel are immediately cancelled as they likely
will not be relevant in the future if the channel is ever reactivated.

task-3547088

X-original-commit: f229baa9fe98d1f10d2674b8cdd4dd9eca0645ca
Part-of: odoo/enterprise#85176
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>